backend/requirements.txt
# COMENTARIO DE ARCHIVO - GRAPHHIRED
# Dependencias Python del backend. Partes: framework API, ORM/base de datos, seguridad, IA/LangGraph, parsing de PDF y utilidades.

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# AI/LLM
langgraph==0.2.60
langchain-core==0.3.83
langchain-google-genai==2.1.4

# Database
sqlalchemy==2.0.47
psycopg2-binary==2.9.10
pgvector==0.2.4
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.13.1

# Data Validation
pydantic==2.12.5
pydantic-settings==2.11.0

# File Processing
python-multipart==0.0.9
pypdf2==3.0.1
python-docx==1.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bleach==6.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
tenacity==8.2.3

email-validator